chore(collaborator hooks): shift into own files for ease of discovery
